don't loop infinitely if a typo spans multiple tokens
This happens because the "default" token gets emitted for each not matched
char resulting in lots of small "default" tokens.

This markdown:

> typu

will result in the token stream
{"default", 2, "default", 3, "default", 4, "default", 5}
and a typo stream with a single entry spanning all of the tokens
{{"typu", 1, 5}}.

Not checking if the typo is longer then the current token leads to an infinite
loop where the typo is always inserted in the new token stream without advancing
the token stream.