fix typo detection at file start
If a file starts with a typo our pattern [%A]typo[%A] can not find it.
However if the same typo appears a second type in the text our pattern
happily finds the second occurence skipping the first and all typos
in between.

In the text

typo1
Correct text typo2 ...
typo1

our pattern will match the second occurence of 'typo1' skipping
'typo2'.
Since we update our search index 'typo2' will never be found and
vis-spellcheck will show the user an information prompt.

Fix this behaviour by adding a special case for the first typo which may
start the text.